int to uint32_t

+++ src/clienthandlers/SolverClientHandler.hpp
@@ -17,10 +17,10 @@ namespace server {
         template <typename Problem, typename Solution>
         class SolverClientHandler : public ClientHandler {
 
-            static const int s_bufferSize = 1000000;
-            static const double s_version = 1.0;
-            static const double s_emptyResponseLength = 0;
-            static const double s_successStatus = 0;
+            static const uint32_t s_bufferSize = 1000000;
+            static const uint32_t s_version = 1.0;
+            static const uint32_t s_emptyResponseLength = 0;
+            static const uint32_t s_successStatus = 0;
             mutable cache::CacheManager m_cache;
 
             public:
